#55 feat : Nft 구매 성공

+++ app/src/main/java/com/aoztg/greengrim/data/model/request/PurchaseNftRequest.kt
@@ -2,5 +2,5 @@ package com.aoztg.greengrim.data.model.request
 
 data class PurchaseNftRequest(
     val payPassword: String,
-    val marketId: Int
+    val marketId: Long
 )
